'data'

+++ migrations/20260901073250_order-create-column-cod-fee.js
@@ -2,18 +2,24 @@
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.up = async function(knex) {
-  await knex.schema.alterTable("orders", (table) => {
-    table.decimal("cod_fee", 10, 2).notNullable().defaultTo(0);
-  })
+exports.up = async function (knex) {
+  const hasColumn = await knex.schema.hasColumn("orders", "cod_fee");
+  if (!hasColumn) {
+    await knex.schema.alterTable("orders", (table) => {
+      table.decimal("cod_fee", 12, 2).notNullable().defaultTo(0);
+    });
+  }
 };
 
 /**
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = async function(knex) {
+exports.down = async function (knex) {
+  const hasColumn = await knex.schema.hasColumn("orders", "cod_fee");
+  if (hasColumn) {
     await knex.schema.alterTable("orders", (table) => {
-        table.dropColumn("cod_fee");
-      });
+      table.dropColumn("cod_fee");
+    });
+  }
 };
